requirements.txt
assemblyline
assemblyline-v4-service
python-bidi
